requirements.txt
biopython==1.85
black==25.1.0
certifi==2025.8.3
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
distlib==0.4.0
filelock==3.19.1
fonttools==4.59.1
identify==2.6.13
idna==3.10
iniconfig==2.1.0
kiwisolver==1.4.9
markdown-it-py==4.0.0
matplotlib==3.10.5
mdurl==0.1.2
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.3.0
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
typer==0.16.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.34.0
